django/piCourse/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
django-filter==23.5
pytz==2023.3 